+++ GRD/GRD/Controllers/SuppliersController.cs
@@ -147,6 +147,11 @@ namespace GRD.Controllers
                 return Unauthorized();
             }
             var supplier = await _context.Suppliers.FindAsync(id);
+            if (supplier == null)
+            {
+                return View("Views/Suppliers/NotFound.cshtml");
+            }
+            supplier.Products.ToList().ForEach(p => { p.Purchases.ToList().ForEach(pu=>p.Purchases.Remove(pu)); supplier.Products.Remove(p); });
             _context.Suppliers.Remove(supplier);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
