+++ GRD/GRD/Controllers/UsersController.cs
@@ -208,6 +208,11 @@ namespace GRD.Controllers
                 return Unauthorized();
             }
             var user = await _context.Users.FindAsync(id);
+            if (user == null)
+            {
+                return View("Views/Users/NotFound.cshtml");
+            }
+            user.Purchases.ToList().ForEach(p => user.Purchases.Remove(p));
             _context.Users.Remove(user);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
